Move modules in packages
